backend/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
psycopg2==2.9.2
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
